==> analysis/1_acquire_data.Rmd <==
---
title: "Acquire data"
date: "`r Sys.Date()`"
bibliography: references.bib
biblio-style: apalike
---

## About
I will acquire data by downloading the transcripts from each episode of Season 1: Love on the Spectrum and then finding a comparable dating show (potentially Love is Blind) that centers around "Typically Developing" adults in the dating world and downloading those transcripts and looking at differences in language.

### Description
The aim of this script is to acquire data from opensubtitles.org, specifically the transcripts of two different shows (Love on a Spectrum and Love is Blind). Then I will save it to the RStudio Cloud directory, and open it here. 

### Usage
I will directly download the files from opensubtitles.org to my computer, unzip and decompress the files, and create two separate folders: one with the .srt files from Love on the Spectrum and one with the .srt files from Love is Blind. I will then compress these folders into a zip file and upload to directory/data/original/ and then upload the files here. 


<!-- How to run this script: what input it requires and output produced -->

## Setup

```{r setup}
# Script-specific options or packages

```

## Run

<!-- Steps involved in acquiring and organizing the original data -->

View the `data/original/ README.md` file for instructions on how to acquire the data from both television shows.

## Finalize

### Log

<!-- Any description that will be helpful to understand the results of this script and how it contributes to the aims of the project -->

### Session

<details><summary>View session information</summary>

```{r, child="_session-info.Rmd"}
```

</details>

```{r cleanup, echo=FALSE}
rm(list = ls()) # clean working environment
```

## References
