==> analysis/5_generate_article.Rmd <==
---
title: "Do Language Differences Exist Between Adults on the Autism Spectrum and Adults Who Are 'Typically Developed'?"
author: "Asya Little"
date: "`r format(Sys.Date(), '%B %d, %Y')`"
output: 
  bookdown::pdf_document2: 
    number_sections: yes
  bookdown::html_document2: 
    number_sections: yes
    code_folding: hide
bibliography: references.bib
biblio-style: apalike
abstract: | 
  The purpose of this exploratory research was to investigate if differences exist among individuals on the Autism spectrum and those who don't have autism and if so, in what realms. Aware of research on how children with Autism develop and function differently than those who aren't on the spectrum, we test to see if common signs of autism like echolalia and the use of more common and frequently used words exist in adults, especially in a romantic setting. By comparing two dating reality shows on Netflix, *Love on the Spectrum* and *Love is Blind* and investigating their transcripts, we hypothesize that these signs of autism still exist in adults on the spectrum. We found that there is no evidence of echolalia nor the use of more basic and commonly used words in those with autism. 
---

```{r setup, include=FALSE}
# Include options and packages required to generate this report

# Install/Load packages
pacman::p_load(citr, magrittr, knitr, ggplot2)

# Configure default options for knitr code chunks
knitr::opts_chunk$set(fig.width=6, fig.asp = 0.7, fig.align='center', echo=FALSE, message = FALSE, warning = FALSE)
```

# Introduction

<!-- the identification of a relevant topic/ area of interest to you, -->

The purpose of this exploratory research is to investigate if common signs of Autism Spectrum Disorder like echolalia and the use of commonly used words exist in adults on the spectrum as they do in children.

We will investigate two Netflix series, *Love on the Spectrum* and *Love is Blind* and we are interested in looking at whether these common markers of Autism Spectrum Disorder in childhood exist in the conversations (and transcripts of these conversations) of the participants in *Love on the Spectrum*. We will compare their lexicons to see any indications of echolalia, the repetition of the speech of others [@sterponi2014] and the use of more common and basic words in those on the spectrum versus their neurotypical counterparts.

It is important to note that the two shows are not exactly the same. While *Love on the Spectrum* is more of a reality television show that follows the individuals around as they navigate dating and having romantic relationships and is a more natural progression, *Love is Blind* is structured more like a dating (game) show we see nowadays, where there is a reward for "winning" and this seems more forced compared to the natural progression of a relationship with Love on the Spectrum. Nevertheless, we point this out because we did our best to find comparable shows to Love on the Spectrum.

## Background

<!-- a description a viable and thought-provoking research question supported by previous research on the topic in the primary literature, -->

As defined by Lord et al., Autism Spectrum Disorder is defined as "a constellation of early-appearing social communication deficits and repetitive sensory-motor behaviors." [@lord2018]. While we know Autism is a lifelong condition, most studies focus their research on children with autism and comparing them to their neurotypical counterparts. Most individuals on the spectrum are diagnosed as children, therefore the symptoms to diagnose are most seen in children. For example, individuals on the spectrum present deficits in having normal back-and-forth conversation and in developing and maintaining relationships, including romantic relationships. Another common symptom is repetitive speech. Moreover, previous research has shown that adults on the spectrum don't use more than simple phrases and their IQs are in the range of intellectual disability [@lord2018]. It makes sense to believe that signs and symptoms of ASD that occur in childhood are still present and affect the everyday lives of individuals on the spectrum.

In this study, we will focus on two specific signs and symptoms that have been categorized as such of ASD: echolalia and the use of common and simple words/phrases, like stated above. Echolalia is defined as the repetition of the speech of others [@sterponi2014]. This symptom is common in children with Autism, and we will investigate whether this symptom persists into adulthood and how this affects the individual's lexicon, and compare it to neurotypical individuals.

We also wish to investigate the use of simple, common words in those on the Spectrum. Previous research has shown that 50% of its 1,000 participants on the spectrum showed delays in language acquisition by adulthood, so we expect some differences in language use and vocabulary in those on the spectrum compared to neurotypical counterparts [@hus2007],[@lord2018].

## Hypothesis

<!-- a clear, and testable hypothesis based on evidence in the literature, -->

We will take a look at two Netflix series, Love on the Spectrum and Love is Blind in order to investigate if these two signs of ASD exist in adults in a romantic setting. Previous research shows that those with ASD have a difficult time maintaining relationships, especially romantic relationships [@strunz2017]. We know that dating shows exist, we found an interesting way to investigate lexical differences between the two groups of people: a dating show. *Love on the Spectrum* focuses on those on the spectrum and their navigation of dating and obtaining and maintaining romantic relationships and comparing their language and patterns surrounding their language and comparing it to the series *Love is Blind*.

We hypothesize that the transcripts of the episodes from *Love on the Spectrum* will show signs of echolalia as well as the use of simple and commonly used words, compared to the transcripts of the episodes from *Love is Blind*.

# Methods

Data gathering, manipulation, and statistical analysis were conducted with the R Statistical Software [@r2016software].

## Data source

<!-- the identification of an adequate corpus resource (existing or to be developed) to address the research question;  evaluating the strengths and weakness of this particular resource, -->

We obtained our transcripts from opensubtitles.org. Navigate to the `data/original/README.md` markdown item to view the precise steps in manually downloading the transcripts from *Love on the Spectrum* and *Love is Blind*.

## Statistical approach

<!-- reference to and use of the appropriate analytical and/or statistical methods to assess the evidence, -->

This research was exploratory, meaning that we had no set place we desired to end up. Instead, we investigate using different skills learning in the course to apply to our datasets and provide an analysis about what this could mean for this specific data. For this project, we did a Word Frequency Analysis, Type-Token Ratio, Relative Frequency, Topic and Sentiment Modeling to investigate the differences, if any, between individuals on the spectrum in a dating setting and those not on the spectrum in a similar setting.

# Results

Here is example code for a table with a caption in Table \@ref(tab:table1).

```{r table1}
cars %>% 
  head %>% 
  kable(booktabs = TRUE, 
        caption = "Preview of the `cars` dataset.")
```

Here is example code for a figure with a caption Figure \@ref(fig:plot1).

```{r plot1, fig.cap="Plot of the `cars` dataset"}
ggplot(cars, aes(x = dist, y = speed)) + geom_point() + geom_smooth(method = "lm")
```



# Discussion

Our purpose was to investigate whether or not we see differences in word frequency (due to a common characteristic of autism being *echolalia*, the repetition of the speech of others) [@sterponi2014]. Due to this common characteristic, we were hoping to find patterns in the Love on the Spectrum data to validate this hypothesis, however it does not appear that words indicative of repetition (questioning why if someone asks why, etc.) occur in this data.\


We also wished to investigate which words were more common in each dataset. We found that word names seemed to appear most often and when controlling for common words expressed in everyday language use and names so that it wouldn't interfere with us seeing the content of what was talked about in both series, we found words indicative of relationships and romance in *Love is Blind.* Moreover, for *Love on the Spectrum,* we found that quite often, they discussed their Autisim (with the use of words like autism, autistic, spectrum, disability, and asperger's) and discussed more "nerdier" things like palentology, dinosaurs, and manga.


To identify and view trends of whether or not we see echolalia in individuals on the spectrum, we conducted a Type-Token Ratio Analysis. Our plot (which is found in the analyze portion of this website) shows that participants in *Love on the Spectrum* had more diversity in their lexicon, which is contrary to what we hypothesized. Instead, we believed that the individuals on *Love on the Spectrum* used more common and basic words and that there would be more repetition in their transcripts. Instead, we found that the *Love is Blind* has more repetition and *Love on the Spectrum* is more diverse and has more words.


Next, we looked at sentiments and topics surrounding each of the TV Series and we didn't find anything indicative that romance was more of a topic in Love is Blind than in Love on the Spectrum. Instead, the only theme that seemed to appear in the Love on the Spectrum topics were the use of common words. We also see that their word choice across topics is very basic, which is also a characteristic of autism; their word choice compared to typically-developed adults is simpler. For the Love is Blind topics, we found the common words topic like we did in Love on the Spectrum, but we also saw the topic of home, which either shows that individuals talked about their lives growing up, or talked about their lives together. Again, however, we did not find that romance was a topic that was more discussed in Love is Blind than in Love on the Spectrum.

For the Sentiment Analysis, we found that sentiments, overall, were expressed more in *Love is Blind* moreso than in *Love on the Spectrum.* Those on *Love is Blind* expressed sentiments more, and even more so, expressed more negative sentiments (especially anger and disgust). Those on *Love on the Spectrum* expressed emotions quite the same, regardless of if they are negative or positive. This did not confirm our hypothesis in which we expected to possibly see more negative expression for those on the spectrum than their typically developing counterparts who didn't deal with the frustrations of interacting with Autism. Maybe if we would've investigated the word "frustrated", we may've seen this pattern. 

# Conclusion

# References
